Refactor simple transform operator helpers
